Only window position changed, update window crop.
When the window move to the edge of screen, first call setWindowCrop,
if the position of window is changed, second call setPosition.
setWindowCrop will let surfaceflinger dispaly the window, but the
lefttop position is the last postion, this will lead the rightbottom
position reduced, and the reduced rect is display the second layer's
window, so the window flicking.
setPosition let surfaceflinger display the window second, and this
time is right rect.

Solution:  When the window move to the edge of screen, only the
position of window is changed, call setPosition and setWindowCrop.

Cherry-picked from https://android.intel.com/#/c/458039

Conflicts:
	services/core/java/com/android/server/wm/WindowStateAnimator.java

Change-Id: I12f2b380823f869dea8aa1f42b3b8c548d016b95
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-18559
Signed-off-by: Wei, ChangzhiX <changzhix.wei@intel.com>
Reviewed-on: https://android.intel.com:443/458039
Reviewed-on: https://android.intel.com:443/468817