Enable shader cache for applications that shared UID with multiple packages too
Frameworks supported a mature mechanism for shader cache per each package already,
but as a special case such as Settings APP, if there are several packages in this
application which means that multiple packages shared the same SharedUserID with it,
it won't initialize the graphics disk caches, thereby APP like Settings have to
rebuild and relink shader every time during launch, which cause a bad launching
performance, so here to enable the GFX and RS cache initialization for multiple
shared packages case too.

As measured on Sofia LTE this saves 100 ms during
the 'flush drawing commands' on a cold launch of the Settings
application and so improve the launch time by the same.

Next:
    Run below test on multiple platforms with different GFX:
        1. CTS
        2. MTBF
        3. full Perf Regregssion test suites

Category: aosp improvement
Domain: AOSP-Framework-GUI
Origin: internal
Upstream-Candidate: yes

Change-Id: If0f927e3399b775804abf1d9a868887951f471c5
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-19982
Signed-off-by: Shuo Gao <shuo.gao@intel.com>
Reviewed-on: https://android.intel.com:443/476092